feat: 책장 상세정보 페이지에 use client 제거
