+++ app/bookshelves/[id]/page.tsx
@@ -1,134 +1,39 @@
-"use client";
-
-import { useState, useEffect } from "react";
-import { useParams, useRouter } from "next/navigation";
-import { Button } from "@/components/ui/button";
-import { ArrowLeft, Settings, Book } from "lucide-react";
-import { GrayBadge, GreenBadge } from "@/components/atoms/badge/TextLabelBadge";
-import {
-  ShelfUpdateDialog,
-  ShelfUpdateOptions,
-} from "@/components/organisms/popup/ShelfUpdateDialog";
 import { fetchBookshelfDetail } from "@/utils/api/BookshelfApi";
 import { BookPreview } from "@/types/BookPreview";
 import { CommonShelf, MAX_SHELF_BOOK_COUNT } from "@/types/Bookshelf";
 import { findBooksPreview } from "@/utils/api/BookPreviewApi";
-import { ShelfBookListTemplate } from "@/components/templates/ShelfBooksTemplate";
+import { BookshelfDetailTemplate } from "@/components/templates/BookShelfDetailTemplate";
+import PageNotFound from "@/app/not-found";
+import { BookshelfDetailResponseSpec } from "@/types/ApiSpec";
 
-export default function BookshelfDetailPage() {
-  const params = useParams();
-  const router = useRouter();
+export default async function BookshelfDetailPage({
+  params,
+}: {
+  params: Promise<{ id: string }>;
+}) {
+  const shelfId = await (await params).id;
+  if (!shelfId) return <PageNotFound />;
 
-  const [shelf, setShelf] = useState<CommonShelf | null>(null);
-  const [books, setBooks] = useState<BookPreview[]>([]);
-  const [showSettings, setShowSettings] = useState(false);
+  // 책장 정보를 불러옴
+  const response = await fetchBookshelfDetail({
+    shelfId: shelfId,
+    side: "server",
+  });
 
-  useEffect(() => {
-    if (!params.id) {
-      setShelf(null);
-      return;
-    }
+  const shelf: CommonShelf = response;
+  const bookIds: string[] = response.bookIds;
 
-    // 책장 정보를 불러옴
-    fetchBookshelfDetail({
-      shelfId: params.id.toString(),
-      side: "client",
-    }).then((data) => {
-      setShelf(data);
-      if (data.bookIds.length === 0) return;
+  // 책장의 책정보를 불러옴
+  const books: BookPreview[] =
+    bookIds.length === 0
+      ? []
+      : (
+          await findBooksPreview(
+            { bookIds: bookIds, categoryCond: null },
+            { pageNumber: 0, pageSize: MAX_SHELF_BOOK_COUNT },
+            "server"
+          )
+        ).books;
 
-      // 책장의 책정보를 불러옴
-      findBooksPreview(
-        { bookIds: data.bookIds, categoryCond: null },
-        { pageNumber: 0, pageSize: MAX_SHELF_BOOK_COUNT },
-        "client"
-      ).then((response) => {
-        setBooks(response.books);
-      });
-    });
-  }, [params.id]);
-
-  const handleUpdateShelf = (updatedShelf: ShelfUpdateOptions) => {
-    if (!shelf) return;
-
-    setShelf({
-      ...shelf,
-      name: updatedShelf.name.trim(),
-      isPublic: updatedShelf.isPublic,
-      createdAt: new Date().toISOString(),
-    });
-  };
-
-  if (!shelf) {
-    return (
-      <div className="min-h-screen bg-background flex items-center justify-center">
-        <div className="text-center">
-          <Book className="w-16 h-16 mx-auto text-muted-foreground mb-4" />
-          <h2 className="text-xl font-semibold mb-2">
-            책장을 찾을 수 없습니다
-          </h2>
-          <Button onClick={() => router.back()}>돌아가기</Button>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="min-h-screen bg-background">
-      <div className="container mx-auto px-4 py-8">
-        {/* Header */}
-        <div className="flex items-center justify-between mb-8">
-          <div className="flex items-center gap-4">
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={() => router.back()}
-              className="flex items-center gap-2"
-            >
-              <ArrowLeft className="w-4 h-4" />
-              책장 목록
-            </Button>
-            <div>
-              <div className="flex items-center gap-3 mb-2">
-                <h1 className="text-3xl font-bold text-foreground">
-                  {shelf.name}
-                </h1>
-                <div className="flex items-center gap-2">
-                  {shelf.isPublic ? (
-                    <GreenBadge text="공개" />
-                  ) : (
-                    <GrayBadge text="비공개" />
-                  )}
-                </div>
-              </div>
-              <p className="text-muted-foreground">
-                {books.length}/50권의 책이 저장되어 있습니다
-              </p>
-            </div>
-          </div>
-          <Button
-            variant="outline"
-            className="flex items-center gap-2 bg-transparent"
-            onClick={() => setShowSettings(true)}
-          >
-            <Settings className="w-4 h-4" />
-            책장 설정
-          </Button>
-        </div>
-
-        <ShelfBookListTemplate searchResults={books} />
-
-        <ShelfUpdateDialog
-          bookshelf={{
-            id: shelf.id,
-            name: shelf.name,
-            isPublic: shelf.isPublic,
-          }}
-          isOpen={showSettings}
-          onClose={() => setShowSettings(false)}
-          onUpdate={handleUpdateShelf}
-        />
-      </div>
-    </div>
-  );
+  return <BookshelfDetailTemplate initShelf={shelf} initBooks={books} />;
 }
